chore: 데이터형 변환 및 이미지 처리 코드 수정

+++ app/src/main/java/com/codepatissier/keki/src/main/consumer/home/model/HomePostRe.kt
@@ -1,7 +1,7 @@
 package com.codepatissier.keki.src.main.consumer.home.model
 
 data class HomePostRe(
-    val postIdx: Int,
+    val postIdx: Long,
     val storeTitle: String,
     val postImgUrl: String
 )
